requirements.txt
django-jazzmin==2.4.6